+++ src/ptp/ptpd2/ptpd_lib.c
@@ -646,6 +646,7 @@ int ptpd_get_snapshot(struct ptpd_port_context *ptpd, struct ptpd_port_snapshot 
 	snapshot->port.domain_number = ptpd->domainNumber;
 	snapshot->port.slave_only = ptpd->slaveOnly;
 	snapshot->port.master_only = ptpd->masterOnly;
+	snapshot->port.last_sync_ifindex = ptpd->lastSyncIfindex;
 	snapshot->current.servo_outlier_threshold = servo_get_outlier_threshold(&ptpd->servo);
 	snapshot->port.effective_comm_caps = ptpd->effective_comm_caps;